+++ integrations/gitlab/gitlab_integration/utils.py
@@ -1,5 +1,6 @@
 from typing import List
 
+import gitlab
 from gitlab import Gitlab
 from gitlab_integration.gitlab_service import GitlabService
 from loguru import logger
@@ -10,6 +11,23 @@ from port_ocean.exceptions.context import EventContextNotFoundError
 RETRY_TRANSIENT_ERRORS = True
 
 
+def generate_gitlab_client(host: str, token: str) -> Gitlab:
+    try:
+        gitlab_client = Gitlab(
+            host, token, retry_transient_errors=RETRY_TRANSIENT_ERRORS
+        )
+        gitlab_client.auth()
+        logger.info("Successfully authenticated using the provided private token")
+    except gitlab.exceptions.GitlabAuthenticationError:
+        gitlab_client = Gitlab(
+            host, oauth_token=token, retry_transient_errors=RETRY_TRANSIENT_ERRORS
+        )
+        gitlab_client.auth()
+        logger.info("Successfully authenticated using the provided OAuth2.0 token")
+
+    return gitlab_client
+
+
 def get_all_services() -> List[GitlabService]:
     logic_settings = ocean.integration_config
     all_tokens_services = []
@@ -18,11 +36,7 @@ def get_all_services() -> List[GitlabService]:
         f"Creating gitlab clients for {len(logic_settings['token_mapping'])} tokens"
     )
     for token, group_mapping in logic_settings["token_mapping"].items():
-        gitlab_client = Gitlab(
-            logic_settings["gitlab_host"],
-            token,
-            retry_transient_errors=RETRY_TRANSIENT_ERRORS,
-        )
+        gitlab_client = generate_gitlab_client(logic_settings["gitlab_host"], token)
         gitlab_service = GitlabService(
             gitlab_client, logic_settings["app_host"], group_mapping
         )